# openapi.yaml
...
openapi: 3.0.3
info:
  title: siyaKhanya API
  version: "0.1.0"
  description: Starter API for predictions, payback and affordability.
servers:
  - url: http://localhost:8080/api
paths:
...
  /communities/{id}/payback:
    get:
      summary: Get payback & NPV for a chosen system
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
        - name: systemId
          in: query
          schema:
            type: string
      responses:
        "200":
          description: payback object
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaybackResponse'
...
components:
  schemas:
...
    PaybackResponse:
      type: object
      properties:
        system_id:
          type: string
        npv:
          type: number
        payback_years:
          type: number
        yearly_cashflows:
          type: array
          items:
            type: object
            properties:
              year: { type: integer }
              net: { type: number }